Add support for setting sprokit edge default capacity.
The default capacity for edges was unlimited.
which is almost never desired. This change sets a default capacpty
to 10 and supplies a CMake configuration option to change it at
configure time. This only applies to the edge default capacity.
All existing pipeline config file controls for edge capacity will work
as before.